+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## 0.17.0-0.7.2 (upcoming)
 
+* [PLT-2226] Set private repository by default
 * [PLT-2289] Add safe-to-evict annotations in Flux pods
 * [PLT-2305][EKS] Asegurar la creación de la política de red en el namespace calico-system para permitir su salida
+
+### Major changes & deprecations
+
+* Docker registry and Helm repository are configured as `private` by default. They can be configured via `private_registry` and `private_helm_repo` in the cluster `ClusterConfig`
 
 ## 0.17.0-0.7.1 (2025-06-05)